Fix writing dictionary data to parquet
In a chunked approach the partitioned data is sliced into
dictionary vectors where the partition vectors are shared
across chunks. A previous chunk row position is changed and
not applicable to the prior chunk when the dictionary content
is resolved. This results in incorrect data being written
to the partitioned files.

In addition, Arrow does not support NULL values in a
dictionary and throws an exception.
The exception is
NotImplemented: Writing DictionaryArray with null encoded in dictionary type not yet supported

The new Constance Vector addition using Arrow REE encoding
cannot yet be written to Parquet.
The exception is
NotImplemented: Unhandled type for Arrow to Parquet schema conversion: run_end_encoded<run_ends: int32, values: int64>

To solve the issues, the dictionary vector representing
the partitioning is flattened into a FlatVector.
As a result the data is copied to persist it across chunks.

In addition, the constant vector is expanded into a flat vector and
then written to parquet.

The ArrowBridge has a new option structure to
indicate if dictionary and constant vectors should be flattened.