Add integrity check to StringIdMap
This PR adds code to cross check two maps in
the StringIdMap class in case a certain
StringIdLease is not found. The ids map to file names
and vice versa.

Additionally, a new VLOG(1) trace point is added when
a StringIdLease is removed from the StringIdMap.
This mirrors a VLOG(1) entry when a new id for a file
is added to the map in the FileHandle.