Move ScopedContextSaver's restoration of context outside destructor. (#7838)
Summary:

ScopedContextSaver's destructor calls EvalCtx::restore, which among other things, propagates errors from the
new context to the original context.  Propagating errors can involve growing the error vector, which entails
allocating memory from the MemoryPool.

Unfortunately, there are a variety of ways this could lead to exceptions, we could run out of memory, we could
enter arbitration at a time when the Driver is being aborted, etc.

Destructors are by default noexcept, so any exception thrown in the destructor causes the binary to crash.

We could set the destructor to noexcept(false), but we could still see crashes, if the ScopedContextSaver is
destructed as part of unwinding the stack.  E.g. if an out of memory exception gets thrown and
ScopedContextSaver's destructor throws another out of memory exception while destructing as part of stack
unwinding.

To address this, I think we need to change ScopedContextSaver to no longer be "Scoped" and instead rely on
explicitly restoring the context when the new context is no longer needed, rather than doing so in the
destructor.

This is admittedly more error prone, but everywhere where we're using it today the only way to exit the scope
is to reach the end of the block (no premature returns, breaks, etc.) or due to an exception, so it's fairly easy to
get this right, at least for now.  We also lose the restoration of the original context (unless we catch, restore,
and rethrow), but it looks like the scopes where ScopedContextSaver is used, if an exception propagates to
that point, the query will fail anyway.

Differential Revision: D51760874