Make ByteStream::size() consistent across use cases
ByteStream::size is expected to have the semantics of the length of
std::stringstream. It is the offset of the first unwritten byte from
the start. This does not move backward when seeking to mid stream and
writing, for example.  been written.

Adds a test to verify proper function when randomly overwriting a
ByteStream backed by HashStringAllocator multipart entries.