Copy input in json_parse to avoid ASAN error (#7858)
Summary:

`value` is put on offset 32 of the stack frame, and the total frame size is 88.  In case `value` is inlined, and SIMD register is 64 bytes (e.g. AVX512), simdjson is reading the memory from offset 36 to 100, which exceeds the frame boundary.

This caused some ASAN error, but no issue in production, because the padding data is not really used or changed.

Fix it by asking `simdjson::dom::parser::parse` to copy the input to internal padded memory owned by parser.  There is some minor performance loss but is probably not noticeable at query level.

Reviewed By: mbasmanova

Differential Revision: D51810542