Avoid reading underlying stream when seeking RLE decoders
Summary:
`seekToRowGroup` is meant to be a lightweight operation and should not
read the underlying stream.  Currently this is not the case with RLE decoders,
because we eagerly skip the values after seeking, which requires read the RLE
headers from the encoded data.  We have pattern that seek multiple times before
actually reading the data, and the IO and decompression cost of seeking can be
avoided.  This is especially visible in low selectivity queries where
decompression takes majority of the CPU time.

Optimize these repeated IO and decompression away by lazily skipping values in
RLE decoders.  For a typical low selectivity query, this improve the CPU time
from 69.84 hours to 40.2 hours (Presto Java is taking 41.45 hours on the same
query).

Differential Revision: D51947121