Fix parallel build overflow for last bucket
A parallel build inserts one range of buckets per thread. If an insert does not fit in the last bucket in the range, it is added to overflows. The overflows are inserted sequentially at the end of the build.
When inserting overflows, htere are no partition bounds and as long as there is at least one free slot the insert cannot fail.

However, when inserting the overflows, the upper bound of the
partition must be -1 to indicate no bounds. If it is sizeMask + 1 and
the last bucket is full, the insert cannot wrap around to the first
bucket like it should.