Fix race condition between task terminate and resume
There is a race condition between task error and task resume which can lead to
zombie task. It is caught by arbitrator concurrent test
T1. task request pause
T2. task fails by external event
T3. one driver tries to get on thread but found the task has been terminated
and stopped. Task::shouldStop return terminate error
T4. the driver fails to get on thread and just return without close the driver as the
task is terminated so it assumes that the task terminate process will close the
driver.
However, the task terminate doesn't close off thread drivers if the task is paused,
and it let's the task resume close it. This cause the pending reference on the task
from the unclosed driver.
This PR fixes the race condition by checking the task stop state first in Task::shouldStop.
Verifies with a dedicated unit test and concurrent test.