Add serialization API for set of rows
Top level rows are more efficiently serialized in row sets rather than
arrays or ranges. Arrays of ranges are still useful for repeated
nested content.  The row set path can uses SIMD to gather nulls and
extract idices of non-null values for serialization.

A Scratch objett is added to signatures to pass reusable scratch
memory also for top level calls to range serializing serializatin
functions. This can remove malloc use for temporary vectors.

The API is tested standalone but is not connected to running code, so
this diff does not affect running systens.