Fix the race between memory arbitration and test finish
The background memory arbitration might fail a task after all the results have been
consumed by readCursor and before the task finishes. So when readCursor wait for
task to finish, the state transition might time out if it happens as the task state is a
failure state.
This PR fixes the race in readCursor by checking if the task state is a failure state when
the state transition time out, and throw the task error instead of return a task finish wait
timeout failure back to the test. We found this in memory arbitration concurrent test
which expect memory related failures from test. Verified with 1'000 test iterations.