Prevent infinite looping from hash table probe operations
There are infinite loops in hash table probe operations that
break until find the search entry or find out an empty slot.
This could result in infinite loop when there are some bugs in
the relevant code path. To prevent this, we shall break out of
the loop if we have searched through all the buckets in the
table