Add basic WindowFuzzerTest (#7917)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/7917

This diff adds a basic version of window fuzzer test that tests the Window operator with window functions and aggregation functions.

The window fuzzer test currently have the following limitations:
* Not supporting reproduction of test failures from saved files.
* Not using randomly-generated frames.
* Not allowing custom result verifiers.
* Verify results against DuckDB instead of Presto.

The limitations will be addressed next in the subsequent PRs.

This is the second piece of https://github.com/facebookincubator/velox/issues/7754.

Differential Revision: https://internalfb.com/D51880255

fbshipit-source-id: 4e03cf0020151f9bfe96d1d27182ad3b27f70b24