Replace the overuse of StringView with std::string_view in Re2Functions.h/cpp
Currently there are some overuse of StringView in Re2Functions.h/cpp. It's not
good because StringView is not a replacement of std::string_view, its copy is
expensive than std::string_view, its data() is not safe to point to for copied
short StringView.

In this PR we replace StringView with std::string_view if it is a field of struct
or a function param which is not feed by XxxVector directly.