Fix variance edge case (#8028)
Summary:

When the current accumulator is empty, and then we are adding another
accumulator with `m2` is infinity and `mean` with a very large value that when
squared would become infinity, the new `m2` calculation is indeterminate ($` \infty \times 0`$) and becomes `NaN`.  The correct new `m2` should be infinity.  We should just
take the values of second accumulator directly in this case.

Also update the merge function of covariance just to be safe.

Differential Revision: D52132799