Add per-driver cpu time slicing support
Support per-driver cpu time slicing. It provides an advisory method
from driver to check if a driver has been running too long on the
thread. The driver execution framework respect this limit by checking
this when switch operators in the driver pipeline execution. The long
running operator method shall also respect this limit by checking this
inside the operator method execution such as table scan and spill read
path in hash join.

Add one query config: driver_cpu_time_slice_limit_ms to allow query
system to configure on query execution, and adds metric to monitor the
cpu yield events in production: driver_yield_count

Unit test is added to verify.

The followup is to add a system-wide config in Prestissimo to configure
this for slow query which are blocked in the queue for too long.