Revert D51867634: Multisect successfully blamed "D51867634: [velox][PR] Scatter struct nulls when deserializing Presto wire format" for otest failure (#8105)
Summary:

This diff is reverting D51867634
D51867634: [velox][PR] Scatter struct nulls when deserializing Presto wire format by oerling has been identified to be causing the following test failure:

Tests affected:
- [xldb/alpha_jni/alpha-spark:alphajni_lib_test-java11 - testRleEncodingTopLevelNulls(com.facebook.xldb.presto.shaded.com.facebook.presto.common.type.Type)[16 : Type(row(field_0 integer,field_1 integer))] (com.facebook.xldb.alpha.EndToEndTest)](https://www.internalfb.com/intern/test/281475090996697/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3818385
Here are the tasks that are relevant to this breakage:


We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D52278809