Ensure wrapping over a loaded lazy vector strips the lazy layer
Currently, dictionary or constant vectors remain in uninitialized
state if they are wrapped over an unloaded lazy vector (lets call
it valuesVector). However, that state needs to be initialized
correctly if the valuesVector either loaded explicitly (not via
the wrapped layer) or the wrapping was done over a loaded lazy
vector. In both these cases, the dictionary or constant would end
up pointing to a loaded lazy vector which would retain the lazy
layer. This is wasteful and can cause issues during expression
eval where peeling would only peel away the encoding layer and
pass on the lazy layer to udfs and expression.
This change ensure that the lazy layer is stripped away by
pointing the vector returned by valuesVector->loadedVectorShared().
This also the ensures internal state of valuesVector is always
initialized incase it has nested encoding layers.

Test Plan:
Added test cases in VectorTest