Fix SsdFileTest
initializeCache() in SsdFileTest used to call MemoryAllocator::
createDefaultInstance() to get the static global MemoryAllocator.
However, it would find customInstance_ not null but an object of
the abstract class MemoryAllocator which does not have the
implementation for the pure virtual functions, and thus causes
"pure virtual method called" error. This commit fixes it by
resetting the allocator first, then explicitly create an
MmapMemoryAllocator which inherits MemoryAllocator and have all
virtual functions implemented.