Guard against AppendWindow destructor throw
AppendWindow is a scoped guard over a fixed size append to a
ByteStream. If the ByteStream does not have space, then a scratch
piece of memory is used to provide contiguous space and is then
appended to the ByteStream on exit. This last append may however
fail. This may throw from stack unwinding and kill the process.

To prevent this, we ensure that all the memory is available at
AppendWindow construction time. The destructor cannot throw because of
failure to extend the stream. We add a LOG(FATAL) to mark the
unreachable throw from destructor.