Fix LocalExchangeSource timeout
Timeout can happen concurrently with the data callback from the output
buffer manager. Only one of these must happen. These are not naturally
serialized because these do not realize the same promise and do all
kinds of other actions in addition to the promise.

Overlap of timeout and data makes exchange fuzzer flaky.