Bound the size of Expr's sharedSubexprResults_ to prevent OOMs (#8128)
Summary:

We recently discovered cases where lambdas with shared subexpressions can cause the size of sharedSubexprResults_ to 
explode by getting invoked with different inputs on each iteration within a batch.  See 
https://github.com/facebookincubator/velox/issues/8093 for a more thorough description.

To avoid this case, we bound the size of sharedSubexprResults_.

I set the default limit to 10 as this seems like a reasonable limit that should work in almost all cases (see the explanation in 
the code comments).  However, it's always possible to imagine cases where this limit wouldn't be high enough, so I made it
configurable.  Note that the only negative effect of having too low a limit would be unnecessarily re-evaluating some shared 
expression trees.

Reviewed By: spershin

Differential Revision: D52335277