Update TaskTest to verify executionEndTimeMs and terminationTimeMs (#7546)
Summary:

Today, we don't have any tests that executionEndTimeMs and terminationTimeMs are not set while the task is
running and set once the task terminates.

This change updates the tests to cover these stats.

In addition, it moves where we set terminationTimeMs to the code in terminate that is only called once, so we
don't need the if statement.

Reviewed By: xiaoxmeng

Differential Revision: D51270201