Allow offline pre-compilation of 'xsimd' dependency
`set(xsimd_SOURCE BUNDLED)` will force the 'xsimd' library to be
compiled from source, but this will make it impossible to compile Velox
offline. We should use `set_source(xsimd)` so that it respects the
`VELOX_DEPENDENCY_SOURCE` global cmake variable.

If it is really needed to force the 'xsimd' library to be compiled from
source in a certain environment, use the following method to achieve it.

```
cmake -Dxsimd_SOURCE=BUNDLED ...
```