remove FOLLY_ALWAYS_INLINE from expensive and test functions
Summary:
We use FOLLY_ALWAYS_INLINE in simple functions when the function cost is too small that a function call 
cost is significant, or when there is potential auto SIMD by the compiler or loop optimizations. 

For complex functions operating on maps and arrays thats almost never the case (except the cardinality function). 
this diff clean up some of the usage.

Differential Revision: D52672969