Add option to limit input size in expression fuzzer (#8315)
Summary:

Add the option maxInputsThreshold to ExpressionFuzzer::options
when maxInputsThreshold is true, if the input size is reached to maxInputsThreshold, then 
when fuzzing new input columns, if an already fuzzed column of the same type exists then it will be used
instead of fuzzing new one. 


This is needed for some internal meta fuzzer tests where input size used to blow up.

Reviewed By: kevinwilfong

Differential Revision: D52642521