Add basic WindowFuzzerTest (#7917)
Summary:

This diff adds a basic version of window fuzzer test that tests the Window operator with window functions and aggregation functions.

The window fuzzer test currently have the following limitations:
* Not supporting reproduction of test failures from saved files.
* Not using randomly-generated frames.
* Not allowing custom result verifiers.

The limitations will be addressed next in the subsequent PRs.

This is the second piece of https://github.com/facebookincubator/velox/issues/7754.

Differential Revision: D51880255