Assert addressable range in bits::scatterBits
If the size in bits is a multiple of 8, scatterBits will read/write
one byte past the last byte in range in target and mask. This can hit
unmapped memory in rare cases, for example if std::vector is used as
container and we have a power of two size and the array ends at the
last byte of mapped address space.

We intentionally reference one byte past end with ASAN, so that ASAN
detects the overread even if the array does not border unmapped
memory.