Preserve constant encodings in Presto serializer
When appending constants, we record the constant value and count of
repeats. We keep a vector that holds a copy of each constant. if it is
all the same constant, we produce a constant. If there are several, we
produce a dictionary vector. We revert to flat vector if flat vectors
are appended or if nulls are appended from null-introducing
dictionaries. We will later consider preserving dictionaries and/or
dictionarizing low cardinality flat vectors.

The test encodedConcatenation goes over appending different constants,
flats and dictionaries into a serializer and reads the data back. The
data is read back from separately encoded pieces to test combining
encodings in deserialization. At the same time, it appends the encoded
pieces into a single serializer to test combining encodings at send
time.