Scatter struct nulls when deserializing Presto wire format
When reading spill serialization, struct nulls are written before the struct columns and the reading can proceed i a single pass.

Like this, nulls from enclosing structs are passed down when
reading. These are combined ith nulls of the contained column so that
the contained column also has a null for rows where the enclosing
struct is null.

When reading Presto Pages, struct nulls come after the child
columns. A separate pass scatters the child column values so as to
create a null gap for the rows where the containing struct is null.

Adds a test for encoding preserving roud trips. Adds a test for
concatenating different encodings in a message, e.g. constant,
dictionary, flat in all combinations of same/different encoding/value
domain. This functionality only applies to nulls first
representations. This will apply to Presto pages when the struct nulls
are read before constructing the struct. See PR 8152 for the end
state.