Fix bucket counts in histogram metrics
Summary:
I noticed Velox's histogram metrics were consuming ~1 GB of memory.

Looking at the comments on the metrics, the authors tried to create the metrics with a small number of buckets, but the argument where they put the number of buckets is actually the width of buckets, so these metrics frequently have tens of thousands of buckets.

This fixes those metrics setting the bucket width so that the number of buckets created matches the comments.

Differential Revision: D52927472