Preserve recursive encodings while serializing PrestoPage (#8249)
Summary:

In the current PrestoPage serialization code, only the first level of
encodings (dictionaries and constants) are preserved, since at times 
VectoStreams are created without the actual data - so that the 
encoding is unknown. 

When data is serialized using `serializeEncoded()` and hence the vector data 
is available, we can recursively capture the encoding and create the appropriate 
streams to serialize dictionaries and constants, thus preventing us from 
flattening them.

Reviewed By: kparichay, mbasmanova

Differential Revision: D52524708