Fix json_parse for large number (#8455)
Summary:

When the number is well-formated but outside the range of int64, we
still allow such cases in `json_parse` so that we can cast them to floating
point numbers later.  This is achieved by replacing the DOM parser with a tree
walk visitor with on-demand parser.  Also relax the corresponding number parsing
in JSON cast.

The `simdjson` version is upgraded in order to use the `simdjson::ondemand::document::at_end` method.  The new version matches the version used inside Meta.

Fix https://github.com/facebookincubator/velox/issues/8066

Reviewed By: mbasmanova

Differential Revision: D52888389