Signal timeout when locking task in getLongRunningOpCalls
Summary:
If there are some problem (e.g. deadlock) with the mutex in `Task`,
`getLongRunningOpCalls` would block forever waiting for it.  In this case we
should signal the error and let the caller to generate some form of alerts.
This is important as we rely on `getLongRunningOpCalls` to detect blocking and
the function itself should not block.

Differential Revision: D53048088