Make copies of Vectors in Values Operator (#8613)
Summary:

The Values Operator takes a std::vector of RowVectors and simply returns them, it's usually used as a source in test queries.

The std::vector of RowVectors comes from the Plan.  There's a risk of concurrency issues because the std::vector is shared 
across Drivers running on multiple threads.

This is demonstrated by the AssertQueryBuilderTest.concurrency test which currently fails with TSAN. It makes a copy of the
output RowVectors, which are simply the Vectors read from Values, which leads to data races as the field 
containsLazyNotLoaded_ gets updated by the concurrent copies.

This also fixes ValuesTest.valuesWithRepeatAndParallelism when run with TSAN enabled.

Since the Drivers are created on a single thread doing the copies when the Operators are created is thread safe.  Since it's
only done once at Driver startup the cost of the copy is minimal.

Differential Revision: D53251319