Allow serializeEncode to accept ranges from callers (#8634)
Summary:

Right now, the serializeEncoded path in the PrestoSerializer does not allow for the caller to specify the ranges of rows in their call to serialize.

This is a departure from the previous behavior where creating a serializer and using ->append() allowed for specifying ranges. 

This diff adds that functionality back, but leaves the option for the caller to not specify ranges, and uses a default full set of ranges if so.

Differential Revision: D53276501