Reduce the lock contention in MallocAllocator by sharded memory reservation (#8569)
Summary:
Reduce the lock contention in MallocAllocator by adding the sharded counters to
reserve a small chunk of capacity from the global allocation counter. When we do
small memory allocations, instead of checking capacity against the global allocation
counter, we check against one of the sharded counter based on the allocation thread
id. When there is insufficient memory reservation in the sharded counter, we reserve
a chunk of capacity from the global counter. When we free small memory allocation, 
we return the memory capacity to one of the sharded counter. If there is too many free
reservation in the sharded counter, we return a chunk of free capacity back to the
global counter. This can help to avoid updating the global allocation counter for every
single memory allocation. This reduces the execution time of concurrent benchmark by
4x.

We introduce an option in MemoryManagerOptions: allocationSizeThresholdWithReservation
to define the memory allocation size threshold to apply this optimization. For memory
allocation larger than this size, we update from the global counter directly. If this is set
to zero, we disable this sharded memory reservation optimization.

This PR also adds ConcurrentCounter abstraction to handle similar use cases in Velox.


Reviewed By: mbasmanova

Differential Revision: D53148151

Pulled By: xiaoxmeng