Ubuntu: build fmt from source in setup script
The default libfmt-dev package installed on
Ubuntu 22.04 is version 8.1.1
(libfmt-dev/jammy,now 8.1.1+ds1-2 amd64 [installed].

However, the fmt version installed needs to be
at least fmt 9.

On CentOS fmt 10.1.1 is downloaded and installed.
This PR now does the same for Ubuntu.

Otherwise, both fmt and folly need to be set to build
from the bundled versions.