Move velox `toTypeSignature` method to a common location (#8695)
Summary:

Moving a shareable `toTypeSignature` method to velox codebase from data_preproc/common. 

The method basically converts a Velox `type` to `typeSignature`

Unit tests:
The tests are split into 3 categories:
* Physical Types
* Logical Types
* Complex Types

For physical and logical types, we've ensured that the signature only has a valid value for the `baseName`.

For complex types, we are ensuring that the base names are correct and as we go down the containers, the values are correct in relevant levels. Skipped some of the validations in the complex type to make sure the code is still readable.

Differential Revision: D53458873