Fix incorrect result of first_value/last_value functions (#8626)
Summary:

The first_value and last_value functions produce incorrect results when used with
IGNORE NULLS and when the frame doesn't start from the first row. In this situation,
FirstLastValueFunction::setRowNumbersIgnoreNulls() attemps to find the index of
the first non-null value from `leastFrame` to `leastFrame + frameSize`. But
the index found is relative to `leastFrame`, so setRowNumbersIgnoreNulls() should
add leastFrame to the index before returning it.

This diff fixes https://github.com/facebookincubator/velox/issues/8427.

Differential Revision: D53295212