Fix arbitrary() to always return the first non-null value in Window operation (#8640)
Summary:

The implementation of arbitrary() intends to always return the first non-null value, but it
doesn't for complex-typed inputs when used in Window operation with incremental frames.
This is because ArbitraryFunction::addSingleGroupRawInput() still updates the accumulator
even if the accumulator already has a value. This diff fixes this issue by making
ArbitraryFunction::addSingleGroupRawInput() return immediately if accumulator already has
a value.

This diff fixes https://github.com/facebookincubator/velox/issues/8593.

Reviewed By: kgpai

Differential Revision: D53328253